# swagger.json
{
   "swagger":"2.0",
   "info":{
      "description":"\"Api de gestão de simulados\"",
      "version":"1.0.0",
      "title":"Teste técnico – Desenvolvedor Backend - SAS",
      "contact":{
         "name":"Gustavo Cabrera Cirilo",
         "url":"https://www.linkedin.com/in/gustavo-cabrera-2894b2a3/",
         "email":"gcabrerac@live.com"
      }
   },
   "host":"localhost:9000",
   "basePath":"/",
   "tags":[
      {
         "name":"prova-controller",
         "description":"Prova Controller"
      },
      {
         "name":"ranking-controller",
         "description":"Ranking Controller"
      },
      {
         "name":"simulado-controller",
         "description":"Simulado Controller"
      }
   ],
   "paths":{
      "/teste-sas/provas/{id_prova}":{
         "get":{
            "tags":[
               "prova-controller"
            ],
            "summary":"listarGabarito",
            "operationId":"listarGabaritoUsingGET",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id_prova",
                  "in":"path",
                  "description":"id_prova",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/GabaritoDTO"
                  }
               }
            },
            "deprecated":false
         }
      },
      "/teste-sas/provas/{id_prova}/responder-questao":{
         "post":{
            "tags":[
               "prova-controller"
            ],
            "summary":"responderQuestao",
            "operationId":"responderQuestaoUsingPOST",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id_prova",
                  "in":"path",
                  "description":"id_prova",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               },
               {
                  "in":"body",
                  "name":"respostaInDto",
                  "description":"respostaInDto",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/RespostaInDTO"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/RespostaOutDTO"
                  }
               },
               "400":{
                  "description":"Bad Request",
                  "schema":{
                     "$ref":"#/definitions/ErrorResponse"
                  }
               },
               "422":{
                  "description":"Unprocessable Entity",
                  "schema":{
                     "$ref":"#/definitions/ErrorResponse"
                  }
               },
               "500":{
                  "description":"Internal Server Error",
                  "schema":{
                     "$ref":"#/definitions/ErrorResponse"
                  }
               }
            },
            "deprecated":false
         }
      },
      "/teste-sas/ranking":{
         "get":{
            "tags":[
               "ranking-controller"
            ],
            "summary":"listarRanking",
            "operationId":"listarRankingUsingGET",
            "produces":[
               "*/*"
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/RankingDTO"
                  }
               }
            },
            "deprecated":false
         }
      },
      "/teste-sas/simulados":{
         "get":{
            "tags":[
               "simulado-controller"
            ],
            "summary":"listarSimulados",
            "operationId":"listarSimuladosUsingGET",
            "produces":[
               "*/*"
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "type":"array",
                     "items":{
                        "$ref":"#/definitions/SimuladoDTO"
                     }
                  }
               }
            },
            "deprecated":false
         }
      },
      "/teste-sas/simulados/{id_simulado}":{
         "get":{
            "tags":[
               "simulado-controller"
            ],
            "summary":"listarSimulado",
            "operationId":"listarSimuladoUsingGET",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id_simulado",
                  "in":"path",
                  "description":"id_simulado",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/SimuladoDTO"
                  }
               }
            },
            "deprecated":false
         }
      }
   },
   "definitions":{
      "Character":{
         "type":"object",
         "title":"Character"
      },
      "ErrorResponse":{
         "type":"object",
         "properties":{
            "errors":{
               "type":"array",
               "items":{
                  "$ref":"#/definitions/FieldValidationMessageResponse"
               }
            },
            "message":{
               "type":"string"
            },
            "status":{
               "type":"integer",
               "format":"int32"
            },
            "timestamp":{
               "type":"string",
               "format":"date-time"
            }
         },
         "title":"ErrorResponse"
      },
      "FieldValidationMessageResponse":{
         "type":"object",
         "properties":{
            "field":{
               "type":"string"
            },
            "message":{
               "type":"string"
            }
         },
         "title":"FieldValidationMessageResponse"
      },
      "GabaritoDTO":{
         "type":"object",
         "properties":{
            "gabarito":{
               "type":"array",
               "items":{
                  "$ref":"#/definitions/QuestaoDTO"
               }
            },
            "nomeProva":{
               "type":"string"
            }
         },
         "title":"GabaritoDTO"
      },
      "ProvaDTO":{
         "type":"object",
         "properties":{
            "idProva":{
               "type":"integer",
               "format":"int64"
            },
            "nomeProva":{
               "type":"string"
            }
         },
         "title":"ProvaDTO"
      },
      "QuestaoDTO":{
         "type":"object",
         "properties":{
            "alternativaCorreta":{
               "$ref":"#/definitions/Character"
            },
            "enunciado":{
               "type":"string"
            },
            "nivel":{
               "type":"string"
            },
            "numeroQuestao":{
               "type":"integer",
               "format":"int32"
            }
         },
         "title":"QuestaoDTO"
      },
      "RankingDTO":{
         "type":"object",
         "properties":{
            "dataDeAplicacao":{
               "type":"string",
               "format":"date"
            },
            "nomeSimulado":{
               "type":"string"
            },
            "provas":{
               "type":"array",
               "items":{
                  "$ref":"#/definitions/ProvaDTO"
               }
            },
            "quantidadeDeProvas":{
               "type":"integer",
               "format":"int32"
            }
         },
         "title":"RankingDTO"
      },
      "RespostaInDTO":{
         "type":"object",
         "properties":{
            "alternativaSelecionada":{
               "$ref":"#/definitions/Character"
            },
            "numeroQuestao":{
               "type":"integer",
               "format":"int32"
            }
         },
         "title":"RespostaInDTO"
      },
      "RespostaOutDTO":{
         "type":"object",
         "properties":{
            "alternativaCorreta":{
               "$ref":"#/definitions/Character"
            },
            "alternativaEscolhida":{
               "$ref":"#/definitions/Character"
            },
            "complexidade":{
               "type":"string",
               "enum":[
                  "FÁCIL",
                  "MÉDIA",
                  "DIFÍCIL"
               ]
            },
            "numeroQuestao":{
               "type":"integer",
               "format":"int32"
            }
         },
         "title":"RespostaOutDTO"
      },
      "SimuladoDTO":{
         "type":"object",
         "properties":{
            "dataDeAplicacao":{
               "type":"string",
               "format":"date"
            },
            "nomeSimulado":{
               "type":"string"
            },
            "provas":{
               "type":"array",
               "items":{
                  "$ref":"#/definitions/ProvaDTO"
               }
            },
            "quantidadeDeProvas":{
               "type":"integer",
               "format":"int32"
            }
         },
         "title":"SimuladoDTO"
      }
   }
}